+++ src/daemon/daemon_test.c
@@ -65,7 +65,7 @@ ahc_nothing (void *cls,
              const char *url,
              const char *method,
              const char *version,
-             const char *upload_data, unsigned int *upload_data_size,
+             const char *upload_data, size_t *upload_data_size,
              void **unused)
 {
   return MHD_NO;
